Client Registration(fix): Fully implements the USER command (#3)
