+++ Makefile
@@ -3,7 +3,7 @@ CC		= c++
 CFLAGS	= -Wall -Wextra -Werror -Wshadow -std=c++20 -MMD -MP
 SRC		= src/main.cpp src/server.cpp src/listener.cpp src/client.cpp \
 		  src/error.cpp src/msg_parser.cpp src/registration.cpp \
-		  src/parsing_utils.cpp src/commands.cpp
+		  src/parsing_utils.cpp src/commands.cpp src/numerics.cpp
 OBJS	= $(SRC:.cpp=.o)
 DEPS	= $(SRC:.cpp=.d)
